# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API ProVentas",
        "description": "Documentación API Laravel 12",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/tipodocumentos": {
            "get": {
                "tags": [
                    "TipoDocumento"
                ],
                "summary": "Listar tipos de documento",
                "operationId": "ebd726ac0e832b615a2b6db7316d6fbf",
                "responses": {
                    "200": {
                        "description": "Lista de tipos de documento"
                    }
                }
            },
            "post": {
                "tags": [
                    "TipoDocumento"
                ],
                "summary": "Registrar tipo documento",
                "operationId": "c96f63bad59f9e4447b3ca6418feed29",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "type"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "FACTURA"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Documento fiscal factura"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "VENTA"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tipo documento registrado"
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                }
            }
        },
        "/tipodocumentos/{id}": {
            "get": {
                "tags": [
                    "TipoDocumento"
                ],
                "summary": "Obtener tipo documento",
                "operationId": "d1c150165ef8a8a0e573aae33bc82fdd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del tipo documento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle tipo documento"
                    },
                    "404": {
                        "description": "Tipo documento no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "TipoDocumento"
                ],
                "summary": "Actualizar tipo documento",
                "operationId": "2ab9975fca7d2cfd267c420ae5f592b1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID tipo documento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "type"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "BOLETA"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Documento boleta"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "COMPRA"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tipo documento actualizado"
                    },
                    "404": {
                        "description": "Tipo documento no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "TipoDocumento"
                ],
                "summary": "Eliminar tipo documento",
                "operationId": "34c63228d0bb4520acadc84ab8d95d52",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID tipo documento",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tipo documento eliminado"
                    },
                    "404": {
                        "description": "Tipo documento no encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}